Add files via upload
The excel file contains the ticker, name, sector and industry of each company of the AMEX, NASDAQ, NYSE.
Beware that some values may no longer be relevant when you'll use the database. For example, a company acquired by another will no longer be listed on the stock market and thus the DataReader function won't work.